Consistency changes to PRS functions
