+++ prody/utilities/legacy.py
@@ -489,4 +489,430 @@ def msaeye(msa, unique, turbo):
     elapsed1 = toc1 - tic1
     LOGGER.debug('Elapsed: %4.2fs'%elapsed1)
 
+def showPerturbResponse(**kwargs):
+    """ Plot the PRS matrix with the profiles along the right and bottom.
+
+    If no PRS matrix or profiles are provided, these will be calculated first
+    using the provided options with a provided model (e.g. ANM, GNM or EDA).
+    So as to obtain different sensors and effectors, normMatrix=True by default.
+
+    If atoms are provided then residue numbers can be used from there.
+    *model* and *atoms* must have the same number of atoms. *atoms* must be an
+    :class:`.AtomGroup` instance.
+
+    :arg prs_matrix: a perturbation response matrix
+    :type prs_matrix: :class:`~numpy.array`
+
+    :arg effectiveness: an effectiveness profile from a PRS matrix
+    :type effectiveness: :class:`~numpy.array`
+
+    :arg sensitivity: a sensitivity profile from a PRS matrix
+    :type sensitivity: :class:`~numpy.array`
+
+    :arg model: any object with a calcCovariance method
+        e.g. :class:`.ANM` instance
+    :type model: NMA
+
+    :arg atoms: a :class: `AtomGroup` instance
+    :type atoms: AtomGroup
+
+    :arg returnData: whether to return data for further analysis
+        default is False
+    :type returnData: bool
     
+    :arg percentile: percentile argument for showAtomicMatrix
+    :type percentile: float
+
+    Return values are prs_matrix, effectiveness, sensitivity, ax1, ax2, im, ax3, ax4
+    The PRS matrix, effectiveness and sensitivity will not be returned if provided. 
+    If returnData is False then only the last five objects are returned.
+    """
+
+    import matplotlib.pyplot as plt
+    import matplotlib
+
+    prs_matrix = kwargs.get('prs_matrix')
+    effectiveness = kwargs.get('effectiveness')
+    sensitivity = kwargs.get('sensitivity')
+    model = kwargs.pop('model')
+    atoms = kwargs.get('atoms')
+    returnData = kwargs.pop('returnData', False)
+
+    if atoms is None:
+
+        if prs_matrix is None:
+            if model is None:
+                raise ValueError('Please provide a PRS matrix or model.')
+            else:
+                prs_matrix = calcPerturbResponse(model)
+
+        if effectiveness is None or sensitivity is None:
+            effectiveness, sensitivity = calcPerturbResponseProfiles(prs_matrix)
+        else:
+            returnData = False
+
+        showMatrix_returns = showAtomicMatrix(prs_matrix, effectiveness, sensitivity, **kwargs)
+
+    else:
+        if not isinstance(atoms, AtomGroup) and not isinstance(atoms, Selection):
+            raise TypeError('atoms must be an AtomGroup instance')
+        elif model is not None and atoms.numAtoms() != model.numAtoms():
+            raise ValueError('model and atoms must have the same number atoms')
+
+        if prs_matrix is None: 
+            if model is None:
+                raise ValueError('Please provide a PRS matrix or model.')
+            atoms, prs_matrix = calcPerturbResponse(model=model,atoms=atoms)
+
+        if effectiveness is None or sensitivity is None:
+            effectiveness, sensitivity = calcPerturbResponseProfiles(prs_matrix, atoms=atoms)
+
+        showMatrix_returns = showAtomicMatrix(prs_matrix, effectiveness, sensitivity, **kwargs)
+
+    if not returnData:
+        return showMatrix_returns
+    elif kwargs.get('prs_matrix') is not None:
+       if atoms is not None:
+           return atoms, effectiveness, sensitivity, showMatrix_returns
+       else:
+           return effectiveness, sensitivity, showMatrix_returns
+    else:
+        return prs_matrix, effectiveness, sensitivity, showMatrix_returns
+
+def calcPerturbResponseProfiles(prs_matrix, atoms=None):
+    """ Calculate the effectiveness and sensitivity
+    profiles, which are the averages over the rows
+    and columns of the PRS matrix.
+
+    :arg prs_matrix: a perturbation response matrix
+    :type prs_matrix: ndarray 
+
+    When an *atoms* instance is given, the profiles will be added as data, 
+    which can be retrieved with ``atoms.getData('effectiveness')`` and 
+    ``atoms.getData('sensitivity')``. 
+    """
+
+    effectiveness = np.mean(prs_matrix, axis=1)
+    sensitivity = np.mean(prs_matrix, axis=0)
+
+    if atoms is not None:
+        try:
+            ag = atoms.getAtomGroup()
+            defdata = np.zeros(ag.numAtoms(), dtype=float)
+            ag.setData('effectiveness', defdata.copy())
+            ag.setData('sensitivity', defdata.copy())
+        except AttributeError:
+            pass
+        atoms.setData('effectiveness', effectiveness)
+        atoms.setData('sensitivity', sensitivity)
+        
+    return effectiveness, sensitivity
+
+def showPerturbResponseProfiles(prs_matrix, atoms=None,**kwargs):
+    """Plot as a line graph the average response to perturbation of
+    a particular residue (a row of a perturbation response matrix)
+    or the average effect of perturbation of a particular residue
+    (a column of a normalized perturbation response matrix).
+
+    If no PRS matrix or profiles are provided, these will be calculated first
+    using the provided options with a provided model (e.g. ANM, GNM or EDA).
+    So as to obtain different sensitivity and effectiveness, normMatrix=True by default.
+
+    If no residue number is given then the effectiveness and sensitivity
+    profiles will be plotted instead. These two profiles are also returned
+    as arrays for further analysis if they aren't already provided.
+
+    :arg prs_matrix: a perturbation response matrix
+    :type prs_matrix: ndarray
+
+    :arg atoms: a :class: `AtomGroup` instance for matching 
+        residue numbers and chain IDs. 
+    :type atoms: AtomGroup
+
+    :arg effectiveness: an effectiveness profile from a PRS matrix
+    :type effectiveness: list
+
+    :arg sensitivity: a sensitivity profile from a PRS matrix
+    :type sensitivity: list
+
+    :arg model: any object with a calcCovariance method
+        e.g. :class:`.ANM` instance
+        *model* and *atoms* must have the same number of atoms.
+    :type model: NMA
+
+    :arg chain: chain identifier for the residue of interest
+        default is to make a plot for each chain in the protein
+    :type chain: str
+
+    :arg resnum: residue number for the residue of interest
+    :type resnum: int
+
+    :arg direction: the direction you want to use to read data out
+        of the PRS matrix for plotting: the options are 'effect' or 'response'.
+        Default is 'effect'.
+        A row gives the effect on each residue of peturbing the specified 
+        residue.
+        A column gives the response of the specified residue to perturbing 
+        each residue.
+        If no residue number is provided then this option will be ignored
+    :type direction: str
+
+    :arg returnData: whether to return profiles for further analysis
+        default is False
+    :type returnProfiles: bool
+    """
+    from .perturb import PRSMatrixParseError
+
+    model = kwargs.get('model')
+    if not type(prs_matrix) is np.ndarray:
+        if prs_matrix is None:
+            if model is None:
+                raise ValueError('Please provide a PRS matrix or model.')
+            else:
+                if kwargs.get('normMatrix') is None:
+                    kwargs.set('normMatrix',True)
+                prs_matrix = calcPerturbResponse(**kwargs)
+        else:
+            raise TypeError('Please provide a valid PRS matrix (as array).')
+
+    if atoms is None:
+        raise ValueError('Please provide an AtomGroup object for matching ' \
+                         'residue numbers and chain IDs.')
+    else:
+        if not isinstance(atoms, AtomGroup) and not isinstance(atoms, Selection):
+            raise TypeError('atoms must be an AtomGroup instance')
+        elif model is not None and atoms.numAtoms() != model.numAtoms():
+            raise ValueError('model and atoms must have the same number atoms')
+
+    chain = kwargs.get('chain')
+    hv = atoms.getHierView()
+    chains = []
+    for i in range(len(list(hv))):
+        chainAg = list(hv)[i]
+        chains.append(chainAg.getChids()[0])
+
+    chains = np.array(chains)
+    if chain is None:
+        chain = ''.join(chains)
+
+    resnum = kwargs.get('resnum', None)
+    direction = kwargs.get('direction','effect')
+
+    if resnum is not None: 
+        timesNotFound = 0
+        for n in range(len(chain)):
+            if not chain[n] in chains:
+                raise PRSMatrixParseError('Chain {0} was not found in chains'.format(chain[n]))
+
+            chainNum = int(np.where(chains == chain[n])[0])
+            chainAg = list(hv)[chainNum]
+            if not resnum in chainAg.getResnums():
+                LOGGER.info('A residue with number {0} was not found' \
+                            ' in chain {1}. Continuing to next chain.' \
+                            .format(resnum, chain[n]))
+                timesNotFound += 1
+                continue
+
+        profiles = []
+        for n in range(len(chain)):
+            chainNum = int(np.where(chains == chain[n])[0])
+            i = np.where(atoms.getResnums() == resnum)[0][chainNum-timesNotFound] 
+            if direction is 'effect':
+                profiles.append(prs_matrix[i,:])
+            else:
+                profiles.append(prs_matrix[:,i])
+
+    else:
+        effectiveness = kwargs.get('effectiveness')
+        sensitivity = kwargs.get('sensitivity')
+        if effectiveness is None or sensitivity is None:
+            effectiveness, sensitivity = calcPerturbResponseProfiles(prs_matrix)
+        profiles = [effectiveness, sensitivity]
+
+    for profile in profiles:
+        show = showAtomicLines(profile,atoms=atoms,**kwargs)
+
+    returnData = kwargs.get('returnData',False)
+    if returnData:
+        return show, profiles
+    else:
+        return show
+
+def writePerturbResponsePDB(prs_matrix, pdbIn=None, **kwargs):
+    """ Write the average response to perturbation of
+    a particular residue (a row of a perturbation response matrix)
+    or the average effect of perturbation of a particular residue
+    (a column of a normalized perturbation response matrix)
+    into the b-factor field of a PDB file for visualisation in a
+    molecular graphics program.
+    If no chain is given this will be done for that residue in all chains.
+
+    If no residue number is given then the effectiveness and sensitivity
+    profiles will be written out instead. These two profiles are also returned
+    as arrays for further analysis if they aren't already provided.
+
+    :arg prs_matrix: a perturbation response matrix 
+        or a :class:`.AtomGroup` object with a PRS matrix associated as data
+    :type prs_matrix: array or :class:`.AtomGroup`
+
+    :arg pdbIn: file name for the input PDB file where you would like the PRS
+        data mapped
+    :type pdbIn: str
+
+    :arg pdbOut: a list of file names (enclosed in square
+        brackets) for the output PDB file, default is to append
+        the chain and residue info (name and number) onto the pdbIn stem.
+        The input for pdbOut can also be used as a stem if you enter a 
+        single string enclosed in quotes.
+        If no residue number is supplied, chain is ignored and the default 
+        is to append '_effectiveness' and '_sensitivity' onto the stem.
+    :type pdbOut: list
+
+    :arg chain: chain identifier for the residue of interest, default is all chains
+        If you want to analyse residues in a subset of chains, concatentate them
+        together e.g. 'AC'
+    :type chain: str
+
+    :arg resnum: residue number for the residue of interest
+    :type resnum: int
+
+    :arg direction: the direction you want to use to read data out
+        of the PRS matrix for plotting: the options are 'effect' or 'response'.
+        Default is 'effect'.
+        A row gives the effect on each residue of peturbing the specified 
+        residue.
+        A column gives the response of the specified residue to perturbing 
+        each residue.
+        If no residue number is provided then this option will be ignored
+    :type direction: str
+
+    :arg returnData: whether to return effectiveness and sensitivity for analysis
+        default is False
+    :type returnProfiles: bool
+
+    :arg effectiveness: effectiveness profile
+    :type array
+
+    :arg sensitivity: sensitivity profile
+    :type array
+    """
+
+    if not isinstance(prs_matrix, np.ndarray):
+        try:
+            prs_matrix = prs_matrix.getData('prs_matrix')
+        except:
+            raise TypeError('Please provide a valid PRS matrix in numpy ndarray format.')
+
+    try:
+        fi = open(pdbIn,'r')
+        lines = fi.readlines()
+        fi.close()
+    except:
+        raise PRSMatrixParseError('Please provide a valid file name for the input PDB.')
+ 
+    chain = kwargs.get('chain', None)
+
+    structure = parsePDB(pdbIn,subset='ca')
+    structure.setData('prs_matrix',prs_matrix)
+
+    hv = structure.getHierView()
+    chains = []
+    for i in range(len(list(hv))):
+        chainAg = list(hv)[i]
+        chains.append(chainAg.getChids()[0])
+
+    chains = np.array(chains)
+    if chain is None:
+        chain = ''.join(chains)
+
+    resnum = kwargs.get('resnum', None)
+    pdbOut = kwargs.get('pdbOut', None)
+    if pdbOut is None:
+        out_stem = pdbIn.split('.')[0]
+    elif type(pdbOut) is str:
+        out_stem = pdbOut.split('.')[0]
+        pdbOut = None
+
+    if resnum is None:
+        effectiveness = kwargs.get('effectiveness', None)
+        sensitivity = kwargs.get('sensitivity', None)
+        if effectiveness is None or sensitivity is None:
+            effectiveness = np.mean(prs_matrix, axis=1)
+            sensitivity = np.mean(prs_matrix, axis=0)
+
+        structure.setData('effectiveness', effectiveness)
+        structure.setData('sensitivity', sensitivity)
+
+        file_effs_name = '{0}_effectiveness.pdb'.format(out_stem)
+        file_sens_name = '{0}_sensitivity.pdb'.format(out_stem)
+        fileEffs = open(file_effs_name,'w')
+        fileSens = open(file_sens_name,'w')
+
+        for line in lines:            
+            if line.find('ATOM') != 0 and line.find('HETATM') != 0 and line.find('ANISOU') != 0:
+                fileEffs.write(line)                    
+                fileSens.write(line)
+            elif line.find('ATOM') == 0:
+                fileEffs.write(line[:60] + '{:6.2f}'.format(float(structure.select( \
+                               'chain {0} and resnum {1}'.format(line[21],line[22:26])) \
+                               .getData('effectiveness')) * 100/np.max( \
+                               structure.getData('effectiveness'))) + line[66:])
+                fileSens.write(line[:60] + '{:6.2f}'.format(float(structure.select( \
+                               'chain {0} and resnum {1}'.format(line[21],line[22:26])) \
+                               .getData('sensitivity')) * 100/np.max( \
+                               structure.getData('sensitivity'))) + line[66:])
+            elif line.find('HETATM') == 0:
+                fileEffs.write(line[:60] + '  0.00' + line[66:])
+                fileSens.write(line[:60] + '  0.00' + line[66:])
+                      
+        fileEffs.close()
+        fileSens.close()
+        LOGGER.info('The effectiveness and sensitivity profiles were written' \
+                    ' to {0} and {1}.'.format(file_effs_name,file_sens_name))
+
+        returnData = kwargs.get('returnData',False)
+        if returnData:
+            return structure, effectiveness, sensitivity
+        else:
+            return
+ 
+    direction = kwargs.get('direction','effect')
+    for n in range(len(chain)):
+        if not chain[n] in chains:
+            raise PRSMatrixParseError('Chain {0} was not found in {1}'.format(chain[n], pdbIn))
+
+    if pdbOut is None:
+        pdbOut = []
+        for c in chain:
+            pdbOut.append('{0}_{1}_{2}{3}_{4}.pdb' \
+                          .format(out_stem, c, \
+                                  str(structure.select('chain {0} and resnum {1}' \
+                                      .format(c, resnum)).getResnames()), \
+                                  resnum, direction))
+
+    for c in chain:
+        fo = open(pdbOut[n],'w')
+        for line in lines:
+            if line.find('ATOM') != 0 and line.find('HETATM') != 0 and line.find('ANISOU') != 0:
+                fo.write(line)
+            elif line.find('ATOM') == 0:
+                if direction is 'effect':
+                    fo.write(line[:60] + '{:6.2f}'.format(float(structure.getData('prs_matrix') \
+                                         [structure.select('chain {0} and resnum {1}' \
+                                          .format(c, resnum)).getResindices(), \
+                                          structure.select('chain {0} and resnum {1}' \
+                                          .format(line[21], line[22:26])).getResindices()])*100) \
+                             + line[66:])
+                else:
+                    fo.write(line[:60] + '{:6.2f}'.format(float(structure.getData('prs_matrix') \
+                                         [structure.select('chain {0} and resnum {1}' \
+                                          .format(line[21], line[22:26])).getResindices(), \
+                                          structure.select('chain {0} and resnum {1}' \
+                                          .format(c, resnum)).getResindices()])*100) \
+                             + line[66:])
+            elif line.find('HETATM') == 0:
+                fo.write(line[:60] + '  0.00' + line[66:])
+
+        LOGGER.info('Perturbation responses for specific residues were written' \
+                    ' to {0}.'.format(', '.join(pdbOut)))
+
